reworked the build post process
