+++ src/Sentry.Unity.Editor/Properties/AssemblyInfo.cs
@@ -1,4 +1,6 @@
 using System.Runtime.CompilerServices;
 
 [assembly: InternalsVisibleTo("Sentry.Unity.Editor.iOS")]
+[assembly: InternalsVisibleTo("Sentry.Unity.Editor.iOS.Tests")]
 [assembly: InternalsVisibleTo("Sentry.Unity.Editor.Tests")]
+
